return 12586269025L when calculate given position is 50

+++ src/main/java/cc/xpbootcamp/warmup/fibonacci/Fibonacci.java
@@ -2,7 +2,7 @@ package cc.xpbootcamp.warmup.fibonacci;
 
 public class Fibonacci {
 
-    public int getValueAt(int position) {
+    public long getValueAt(int position) {
         if (position == 1)
             return 1;
         if (position == 2)
